FoFiOutputFunc: Use size_t instead of int

+++ fofi/FoFiBase.h
@@ -13,7 +13,7 @@
 // All changes made under the Poppler project to this file are licensed
 // under GPL version 2 or later
 //
-// Copyright (C) 2018 Albert Astals Cid <aacid@kde.org>
+// Copyright (C) 2018, 2022 Albert Astals Cid <aacid@kde.org>
 // Copyright (C) 2022 Oliver Sander <oliver.sander@tu-dresden.de>
 //
 // To see a description of the changes please see the Changelog file that
@@ -26,7 +26,7 @@
 
 //------------------------------------------------------------------------
 
-using FoFiOutputFunc = void (*)(void *stream, const char *data, int len);
+using FoFiOutputFunc = void (*)(void *stream, const char *data, size_t len);
 
 //------------------------------------------------------------------------
 // FoFiBase
